+++ udata/commands/init.py
@@ -3,8 +3,13 @@ from __future__ import unicode_literals
 
 import logging
 
-from udata.commands import manager, yellow
-from udata.search import es
+from flask_script import prompt_bool
+
+from udata.commands import manager, IS_INTERACTIVE, green
+from udata.core.dataset.commands import licenses
+from udata.core.user import commands as user_commands
+from udata.i18n import lazy_gettext as _
+from udata.search.commands import init as init_search
 
 from .db import migrate
 from .fixtures import generate_fixtures
@@ -14,14 +19,25 @@ log = logging.getLogger(__name__)
 
 @manager.command
 def init():
-    '''Initialize or update data and indexes'''
-    log.info('Initialize or update ElasticSearch mappings')
-    es.initialize()
-
-    log.info('Build sample fixture data')
-    generate_fixtures()
+    '''Initialize your udata instance (search index, user, sample data...)'''
 
     log.info('Apply DB migrations if needed')
     migrate(record=True)
 
-    log.info('%s: Feed initial data if needed', yellow('TODO'))
+    init_search(delete=True, force=not IS_INTERACTIVE)
+
+    if IS_INTERACTIVE:
+        text = _('Do you want to create a superadmin user ?')
+        if prompt_bool(text, default=True):
+            user = user_commands.create()
+            user_commands.set_admin(user.email)
+
+        text = _('Do you want to import some data-related license ?')
+        if prompt_bool(text, default=True):
+            licenses()
+
+        text = _('Do you want to create some sample data ?')
+        if prompt_bool(text, default=True):
+            generate_fixtures()
+
+    log.info(green(_('Your udata instance is ready !')))